requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]
pymongo
certifi
scikit-learn

##-e .
